Implemented JafsInodePool so inode objects can be reused.

+++ src/main/java/nl/v4you/jafs/JafsDirEntry.java
@@ -7,7 +7,7 @@ class JafsDirEntry {
 	
 	long bpos;
 	int ipos;
-	byte type;
+	int type;
 	byte name[];
 	
 	boolean isFile() {
